+++ src/GafferScene/UVSampler.cpp
@@ -34,28 +34,69 @@
 //
 //////////////////////////////////////////////////////////////////////////
 
-#include "boost/python.hpp"
-
-#include "PrimitiveSamplerBinding.h"
-
-#include "GafferScene/ClosestPointSampler.h"
-#include "GafferScene/CurveSampler.h"
-#include "GafferScene/PrimitiveSampler.h"
 #include "GafferScene/UVSampler.h"
 
-#include "GafferBindings/DependencyNodeBinding.h"
-
-using namespace boost::python;
+using namespace Imath;
+using namespace IECore;
+using namespace IECoreScene;
 using namespace Gaffer;
-using namespace GafferBindings;
 using namespace GafferScene;
 
-void GafferSceneModule::bindPrimitiveSampler()
+GAFFER_GRAPHCOMPONENT_DEFINE_TYPE( UVSampler );
+
+size_t UVSampler::g_firstPlugIndex = 0;
+
+UVSampler::UVSampler( const std::string &name )
+	:	PrimitiveSampler( name )
 {
+	storeIndexOfNextChild( g_firstPlugIndex );
 
-	GafferBindings::DependencyNodeClass<GafferScene::PrimitiveSampler>();
-	GafferBindings::DependencyNodeClass<GafferScene::ClosestPointSampler>();
-	GafferBindings::DependencyNodeClass<GafferScene::CurveSampler>();
-	GafferBindings::DependencyNodeClass<GafferScene::UVSampler>();
+	addChild( new StringPlug( "uv", Plug::In, "uv" ) );
+}
 
+UVSampler::~UVSampler()
+{
+}
+
+Gaffer::StringPlug *UVSampler::uvPlug()
+{
+	return getChild<StringPlug>( g_firstPlugIndex );
+}
+
+const Gaffer::StringPlug *UVSampler::uvPlug() const
+{
+	return getChild<StringPlug>( g_firstPlugIndex );
+}
+
+bool UVSampler::affectsSamplingFunction( const Gaffer::Plug *input ) const
+{
+	return PrimitiveSampler::affectsSamplingFunction( input ) || input == uvPlug();
+}
+
+void UVSampler::hashSamplingFunction( IECore::MurmurHash &h ) const
+{
+	PrimitiveSampler::hashSamplingFunction( h );
+	uvPlug()->hash( h );
+}
+
+PrimitiveSampler::SamplingFunction UVSampler::computeSamplingFunction( const IECoreScene::Primitive *primitive, IECoreScene::PrimitiveVariable::Interpolation &interpolation ) const
+{
+	const std::string uv = uvPlug()->getValue();
+	if( uv.empty() )
+	{
+		return SamplingFunction();
+	}
+
+	auto it = primitive->variables.find( uv );
+	if( it == primitive->variables.end() )
+	{
+		throw IECore::Exception( "No primitive variable named \"" + uv + "\"" );
+	}
+
+	interpolation = it->second.interpolation;
+	PrimitiveVariable::IndexedView<V2f> uvView( it->second );
+
+	return [uvView] ( const PrimitiveEvaluator &evaluator, size_t index, const M44f &transform, PrimitiveEvaluator::Result &result ) {
+		return evaluator.pointAtUV( uvView[index], &result );
+	};
 }